LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-compact.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>MacOSInlineMediaControls</code> in compact mode.");

const mediaControls = new MacOSInlineMediaControls({ width: 240 });
mediaControls.layoutTraits = LayoutTraits.macOS | LayoutTraits.Compact;

shouldBeTrue("mediaControls.element.classList.contains('compact')");
shouldBe("mediaControls.leftContainer.leftMargin", "8");
shouldBe("mediaControls.leftContainer.rightMargin", "12");
shouldBe("mediaControls.leftContainer.buttonMargin", "12");
shouldBe("mediaControls.rightContainer.leftMargin", "12");
shouldBe("mediaControls.rightContainer.rightMargin", "8");
shouldBe("mediaControls.rightContainer.buttonMargin", "12");
debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
