LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-compact-expected.txt

Testing MacOSInlineMediaControls in compact mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.classList.contains('compact') is true
PASS mediaControls.leftContainer.leftMargin is 8
PASS mediaControls.leftContainer.rightMargin is 12
PASS mediaControls.leftContainer.buttonMargin is 12
PASS mediaControls.rightContainer.leftMargin is 12
PASS mediaControls.rightContainer.rightMargin is 8
PASS mediaControls.rightContainer.buttonMargin is 12

PASS successfullyParsed is true

TEST COMPLETE
